+++ haero/mam4/validation/CMakeLists.txt
@@ -16,4 +16,4 @@ set(MAM4BOX_BASELINE_DIR ${PROJECT_BINARY_DIR}/mam4box/standalone/tests)
 # Haero Skywalker drivers
 
 add_subdirectory(nucleation)
-#add_subdirectory(gasaerexch)
+add_subdirectory(gasaerexch)
